requirements.txt
# Core Dependencies
pydantic>=2.0.0
asyncio>=3.4.3
aiohttp>=3.8.0
python-dotenv>=0.19.0

# NLP and AI
spacy>=3.5.0
transformers>=4.30.0
torch>=2.0.0
numpy>=1.24.0

# Speech Processing
SpeechRecognition>=3.10.0
pyttsx3>=2.90
sounddevice>=0.4.6
scipy>=1.10.0

# Data Storage
redis>=4.5.0
sqlalchemy>=2.0.0
aiosqlite>=0.17.0

# External Services
google-api-python-client>=2.0.0
google-auth-httplib2>=0.1.0
google-auth-oauthlib>=1.0.0
discord.py>=2.0.0

# Monitoring and Logging
psutil>=5.9.0
prometheus-client>=0.17.0
structlog>=23.1.0

# Testing
pytest>=7.3.1
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
pytest-mock>=3.10.0

# Development
black>=23.3.0
flake8>=6.0.0
mypy>=1.3.0
isort>=5.12.0 